float to double

+++ src/tools.cpp
@@ -21,12 +21,13 @@ VectorXd Tools::CalculateRMSE(const vector<VectorXd> &estimations,
 		cout << "Error - Calculate RMSE () - The groud truth and estimation vector must have same size";
 		return RMSE;
 	}
-	for (uint i = 0; i < estimations.size(); i++)
+	for (unsigned int i = 0; i < estimations.size(); i++)
 	{
 		VectorXd Diff = estimations[i] - ground_truth[i];
 		VectorXd residual = Diff.array() * Diff.array();
 		RMSE = RMSE + residual;
 	}
+	
 	return RMSE;
 }
 
@@ -37,14 +38,14 @@ MatrixXd Tools::CalculateJacobian(const VectorXd& x_state) {
 	MatrixXd Hj(3, 4);
 	//VectorXd h_c(3);
   //Recover state parameters
-	float px = x_state(0);
-	float py = x_state(1);
-	float vx = x_state(2);
-	float vy = x_state(3);
+	double px = x_state(0);
+	double py = x_state(1);
+	double vx = x_state(2);
+	double vy = x_state(3);
 	//Pre-computer a set of terms to avoid repeat calculation
-	float c1 = px * px + py * py;
-	float c2 = sqrt(c1);
-	float c3 = (c1*c2);
+	double c1 = px * px + py * py;
+	double c2 = sqrt(c1);
+	double c3 = (c1*c2);
 	
 	if (c1 < 0.0001) 
 	{
